requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-enumfield==2.0.2
django-filter==2.4.0
django-import-export==2.5.0
django-js-asset==1.2.2
django-resized==0.3.11
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
et-xmlfile==1.0.1
filelock==3.6.0
future==0.18.2
html5lib==1.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
midtransclient==1.2.0
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.6
passlib==1.7.4
Pillow==8.2.0
platformdirs==2.5.1
PyMySQL==1.0.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-dotenv==0.20.0
pytz==2021.1
PyYAML==5.4.1
qrcode==7.3
reportlab==3.5.63
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.5
validate-email==1.3
virtualenv==20.14.1
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
